obs-asio: Switch to Portaudio API
Portaudio has less capabilities than bassasio regarding asio.
But it has the advantage of being compatible with GPL v2+ license.
We have forked portaudio to build with our openAsio sdk which is a drop-in
replacement for regular asio Steinberg sdk but is GPL v2+.
The asioportaudio branch is therefore the one used for releasing binaries.
This version of obs-asio is multi-client but not multi-device :
one can not have several asio sources using different asio devices.
However it is possible to have several asio sources from the same device.
Each client can setup its channels independently so a case use would be
capturing different channel combinations in different tracks.